# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medical-records:
    get:
      operationId: listMedicalRecords
      tags: [medical-records]
      summary: List medical records
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Medical records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecordList"
    post:
      operationId: createMedicalRecord
      tags: [medical-records]
      summary: Create a medical record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
  /medical-records/{id}:
    get:
      operationId: getMedicalRecord
      tags: [medical-records]
      summary: Get medical record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medical record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
    put:
      operationId: updateMedicalRecord
      tags: [medical-records]
      summary: Update medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalRecordInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"

  # ─── Lab Orders ──────────────────────────────────────────────
  /lab-orders:
    get:
      operationId: listLabOrders
      tags: [lab-orders]
      summary: List lab orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lab orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrderList"
    post:
      operationId: createLabOrder
      tags: [lab-orders]
      summary: Create a lab order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"
  /lab-orders/{id}:
    get:
      operationId: getLabOrder
      tags: [lab-orders]
      summary: Get lab order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lab order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"
  /lab-orders/{id}/status:
    patch:
      operationId: updateLabOrderStatus
      tags: [lab-orders]
      summary: Update lab order status/result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabResultUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"

  # ─── Medicines / Pharmacy ────────────────────────────────────
  /medicines:
    get:
      operationId: listMedicines
      tags: [medicines]
      summary: List medicines
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicine"
    post:
      operationId: createMedicine
      tags: [medicines]
      summary: Add a medicine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
  /medicines/{id}:
    get:
      operationId: getMedicine
      tags: [medicines]
      summary: Get medicine by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medicine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    put:
      operationId: updateMedicine
      tags: [medicines]
      summary: Update medicine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    delete:
      operationId: deleteMedicine
      tags: [medicines]
      summary: Delete medicine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    MedicalRecord:
      type: object
      required: [id, patientId, doctorId, visitDate, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        doctorId:
          type: integer
        doctorName:
          type: ["string", "null"]
        visitDate:
          type: string
        chiefComplaint:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        treatment:
          type: ["string", "null"]
        prescription:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        createdAt:
          type: string
    MedicalRecordInput:
      type: object
      required: [patientId, doctorId, visitDate]
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        visitDate:
          type: string
        chiefComplaint:
          type: string
        diagnosis:
          type: string
        treatment:
          type: string
        prescription:
          type: string
        notes:
          type: string
        followUpDate:
          type: string
    MedicalRecordList:
      type: object
      required: [records, total]
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/MedicalRecord"
        total:
          type: integer

    # ─── Lab Orders ────────────────────────────
    LabOrder:
      type: object
      required: [id, patientId, doctorId, testName, status, orderedAt, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        doctorId:
          type: integer
        doctorName:
          type: ["string", "null"]
        testName:
          type: string
        testCategory:
          type: ["string", "null"]
        status:
          type: string
        result:
          type: ["string", "null"]
        referenceRange:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        orderedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    LabOrderInput:
      type: object
      required: [patientId, doctorId, testName]
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        testName:
          type: string
        testCategory:
          type: string
        notes:
          type: string
        orderedAt:
          type: string
    LabResultUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        result:
          type: string
        referenceRange:
          type: string
        notes:
          type: string
        completedAt:
          type: string
    LabOrderList:
      type: object
      required: [labOrders, total]
      properties:
        labOrders:
          type: array
          items:
            $ref: "#/components/schemas/LabOrder"
        total:
          type: integer

    # ─── Medicine ──────────────────────────────
    Medicine:
      type: object
      required: [id, name, category, unit, stockQuantity, price, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        genericName:
          type: ["string", "null"]
        category:
          type: string
        unit:
          type: string
        stockQuantity:
          type: integer
        reorderLevel:
          type: integer
        price:
          type: number
        manufacturer:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    MedicineInput:
      type: object
      required: [name, category, unit, stockQuantity, price]
      properties:
        name:
          type: string
        genericName:
          type: string
        category:
          type: string
        unit:
          type: string
        stockQuantity:
          type: integer
        reorderLevel:
          type: integer
        price:
          type: number
        manufacturer:
          type: string
        expiryDate:
          type: string
        batchNumber:
          type: string
        status:
          type: string
...