# lib/api-zod/src/generated/types/labOrder.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface LabOrder {
  id: number;
  patientId: number;
  /** @nullable */
  patientName?: string | null;
  doctorId: number;
  /** @nullable */
  doctorName?: string | null;
  testName: string;
  /** @nullable */
  testCategory?: string | null;
  status: string;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  referenceRange?: string | null;
  /** @nullable */
  notes?: string | null;
  orderedAt: string;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}