# lib/api-zod/src/generated/types/labOrderList.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { LabOrder } from './labOrder';

export interface LabOrderList {
  labOrders: LabOrder[];
  total: number;
}